fix(nvim): move LuaSnip outside of nvim-cmp
I'm not even sure why but sometimes `<C-l>` mapping does not work after
doing a `PackerCompile` or `PackerSync`. It might have to do something
with the lazy loading feature of packer.nvim

I'm not even sure if this would actually fix the issue but it's
currently working right now so I don't have complaints!