fix(nvim): do not lazy load LuaSnip
Since we use impatient.nvim for caching, then I might actually not need
it. Besides, lazy loading is just deferring plugins to load at a later
time. Sometimes, it causes a small delay when opening a file or
triggering an event because a plugin is being lazy loaded.

Fox example, auto pairs lazy loaded during insert mode causes a slight
delay when going in insert mode for the first time