feat(nvim): switch to fzf-lua
It feels at home coming back to fzf again after months of trying
telescope and snap but I had issues while using both of them.

Telescope has a bug with folds where it doesn't apply folds when opening
files unless you press `zx` to reset the folds.

Snap, on the other hand, was doing great but it was written in fennel
which was quite problematic for me since I was having trouble reading
it. Plus one major issue is with the occasional ENOENT problems while
using the files finder. And also, it lacks tags finder so I had to use
either fzf.vim or telescope to fill that gap.

I think fzf-lua runs an fzf instance which just makes the UI quite nasty
as I don't know yet how to customize the fzf inside of FzfLua but it
works well for me, does its job properly and so I can live with that for
now.