perf(fish): let brew handle system binaries
It has been a big running problem of mine where running asdf +
asdf-direnv and having a global ~/.envrc causes a slowdown to starting
up a fish shell process.

This eliminates that problem where I would just use brew for system
binaries and use asdf+asdf-direnv for project directories. That way, I
would only get punished for the startup time when working inside
projects that require a really specific app version. Else, I could just
use the system binaries as a fallback.

I don't think I'd need something like pyenv. The only reason why I'm
using it was because of a previous experience where the project I'm
working on only workend on python 3.5. I think that's not the case
anymore and projects are starting to upgrade to higher versions of
python.