fix(tmux): set escape-time to 0
It was quite weird since I initially thought karabiner elements was at
fault here as I was setting up a new machine.

It turns out that outside tmux, escape works just fine so this is
probably the culprit. However, I never had this setting when I was using
my MBP 2017.