refactor(nvim): replace null-ls.nvim with conform.nvim + nvim-lint
It's been long overdue since null-ls.nvim posted that it will be
discontinued.

I replaced null-ls.nvim with none-ls.nvim and everything went well.

But none-ls introduced breaking changes that removed most of the
formatting tools I've used by default, so that forced me to move on and
do this refactor.

Closes #6