feat(nvim): add yanky.nvim
yanky is still an experimentation on my end since I haven't used that
feature for a while.

But what it basically does is it allows me improve the existing yank/put
workflow where in I can cycle through each yanked items and allows me to
paste whatever is in my cycle history instead of just the recently
yanked test.