feat(tmux): update split-window bindings for tmux >= 3.4
This commit updates the key bindings for the split-window command in
tmux. The previous bindings were only compatible with tmux versions
less than 3.3a. The new bindings are compatible with tmux version 3.4
and above. The change involves using the '-l' option instead of the '-p'
option to specify the size of the new panes.