fix(nvim): handle nvim-navic require error
This commit modifies the way we require 'nvim-navic' in
'lspconfig.lua'. Instead of directly requiring it, we now use pcall to
catch any errors that might occur during the require process. This
prevents the entire lspconfig setup from failing if 'nvim-navic' is not
available or encounters an error during initialization.