feat(gitconfig): update git aliases and add new ones
This commit updates the 'lg' alias to include more detailed information
in the log. It also moves the 'go' alias to a different location in the
file for better organization. Additionally, a new 'dw' alias is added
for color diffing words. These changes aim to improve the usability and
efficiency of git commands.