feat(nvim): use commentstring_override for htmldjango
There is an existing pull request for
https://github.com/JoosepAlviste/nvim-ts-context-commentstring/pull/92

But the author seems to have ignored it and is not likely to get merged
at some point in the future.

This is just a simple fix for this to just use our custom
commentstring_override which is defined within the
after/ftplugin/htmldjango.lua before even trying to calculate which
commentstring to use.

I tried overriding the vim.bo.commentstring but calculate_commentstring
ignores it.