feat(nvim): add localleader b for buffers
I have not utilized the localleader that much and it feels awkward
pressing `<leader>fb` for [f]ind [b]uffers mnemonic.

`\b` is much easier to press but localleader is quite far.

`,b` could be an option as well but `,` is used for repeating f/t
commands in the opposite direction