feat(nvim): enable exrc
This is quite useful especially when we want to override something by
simply adding .nvim.lua and it will automatically execute its contents.

Previously this was discouraged because it could lead to potential
destructive commands but that's not the case anymore since neovim
provides `:h vim.secure.trust` where it prompts you to allow or deny an
execution.