fix(nvim): do not tint unlisted buffers
I just encountered this when using oil. It has a preview feature when
you press `<C-p>`, it opens a split preview of the file under your
cursor.

But with tint.nvim, it disables the syntax for it. I don't want that to
happen so I'm just gonna ignore this since `nobuflisted` buffers are
just temporary.