perf(nvim): reduce startuptime
From 50ms~ down to 25ms~

The performance slowdown was somewhat caused by which-key, and by adding
`event = VeryLazy` to it, it improved the loading times.

And that's why I made other plugins VeryLazy as well.

For oil.nvim, there was some performance gain at the expense of some
screen flickering when running `nvim .`.

Though I did try to run it without the `VeryLazy` but it seems that it
still flickers. I guess that's okay