feat(nvim): make tabline prettier
I haven't been able to fully utilize having window numbers in the winbar
so it doesn't make sense to set it.

I resorted to just using a nerdicon [cod-mulitiple_windows] because it
just looks appropriate in the given context.

I also made the tabline to only display if there are one or more tabs
just to give an overall nice UI