feat(nvim): add muggle friendly mode
Sometimes when sharing a screen or performing a collab, relativenumbers
just messes things up especially for those that aren't used to it.

I'm not sure where to put this so I just created a new
usercommands.rc.lua for now.

I don't have any other special commands as of the moment.