feat(fish): replace ghq with ghr
When the list of repositories in my ghq source directory got large, it
was getting extremely slow when performing a clone[^1][^2].

It turns out that I have a lot of directories that are not managed by
git especially within my work folder. I was able to figure this out
after moving out the _work_ directory outside the `~/Sources` dir.

I found https://github.com/siketyan/ghr and it seems to be fast because
it doesn't check `.git` folders (which I'm fine with)

Refs:

[^1]: https://github.com/x-motemen/ghq/issues/323
[^2]: https://github.com/x-motemen/ghq/issues/379