feat(yabai): add janky borders
Previously, yabai had borders built in but it got complicated and was
removed from core. janky borders was created to fill the gap to
highlight active border.

Only bug I noticed (as of v1.6.0) is that if you resize applications,
the border only renders half but I can live with that.