feat: replace tint.nvim with sunglasses.nvim
I think sunglasses.nvim is kinda better overall since it provides a
NOSYNTAX filter. I've only been able to emulate that in tint.nvim by
setting a negative value to tint but that still has some colors as it
only makes it gray-ish.

And I've managed to fix the issue where in if you close a tab, the
current window becomes shaded.