refactor(nvim): remove which-key.nvim
I think this is only useful when you aren't using your own distribution.

I'll bring it back if I miss it 😉

But one thing I missed is the spell plugin that which-key provides.

I already created a small wrapper around the builtin spellsuggest in vim
so everything should be just fine.

The built-in spell suggest via `z=` was fine but I find it a bit
intrusive since it occupies a lot of real estate and I have to press
<number> and <CR> when choosing.

fzf-lua provides a spell_suggest function, but I don't quite like it
since you have to press <CR> when choosing a word.