fix(nvim): bring back cursor to last focused window after resisting
I noticed this when I was playing around while creating the spellpopup
plugin.

I was opening a spell suggestion and tried to zoom in my tmux via
<Prefix>z.

I was wondering why the cursor was placed incorrectly. Luckily found
that I had an autocmd to equalize splits when vim is resized.

It was looping through each tab and ran `wincmd =`. The catch was that
it loses focus when setting it to the current tab.