fix(nvim): ensure async behavior when checking for installed servers
I was getting weird issues with fzf because of this. I just noticed it
awhile ago as I was having intermittent internet connection.

My fzf suddenly stops whenever I open a file. It seems that whenever I
open a new file, it first checks the mason registry which might take
awhile if you have slow internet, and that causes opening files to take
a while.

I was using the `mr.refresh` incorrectly. It can be used in 2 ways. If
you pass in a callback, it becomes async.

I was using it as `if mr.refresh()` then `mr.refresh(ensure_installed)`
which was redundant in this case.

I hope this resolves that annoying issue